Fix navigating in results with many matches
Helm seems to have a quirk where it forgets the "real value" of any
candidates that are "below the fold" (i.e. with a line number higher
than the window height). This causes us to receive a display-formatted
string instead of a real value, thus causing "listp" errors.

Work around this by explicitly setting the real value in the
'helm-realvalue property of the display string.

Refactor to allow this and otherwise make the flow more
explicit (e.g. use an explicit helm-git-grep-real-to-display
function).

Bonus fix: move the window-width truncation to where it belongs.